requirements.txt
Django==1.7.1
Pillow
MySQL-python
django-crispy-forms==1.4.0
django-contact-form